add resident set
