+++ device_manager/printer/communicator/communicator.cpp
@@ -11,7 +11,7 @@
 #include "communicator.h"
 #include "../printer.h"
 #include "../printer_manager/printer_manager.h"
-#include "../../printer_utility/printer_utility.h"
+#include "../../../utility/file_utility/file_utility.h"
 
 #define ERROR_STATUS (-1)
