+++ device_manager/printer/communicator/communicator.h
