+++ utility/computer_utility/computer_utility.cpp
@@ -6,7 +6,7 @@
 #include <string>
 #include <cstring>
 #include "computer_utility.h"
-#include "../utility/utility.h"
+#include "../utility.h"
 
 void display_process_continuation_msg(int pid, int base){
     std::string s = "Continuing Execution of {PID->"
@@ -14,14 +14,4 @@ void display_process_continuation_msg(int pid, int base){
                     + std::to_string(base) + "}";
 
     atomically_print_to_stdout(s);
-}
-
-std::vector<int> separate_spaced_str(char str[]){
-    char *word = strtok(str, " ");
-    std::vector<int>a;
-    while(word != nullptr && strcmp(word, "\n") != 0){
-        a.push_back(std::stoi(word));
-        word = strtok(nullptr, " ");
-    }
-    return a;
 }
