+++ utility/file_utility/file_utility.cpp
@@ -5,11 +5,8 @@
 #include <cstdio>
 #include <sys/stat.h>
 #include <unistd.h>
-#include <iomanip>
-#include <iostream>
-#include <sstream>
-#include "printer_utility.h"
-#include "../printer/printer.h"
+#include "file_utility.h"
+#include "../../device_manager/printer/printer.h"
 
 void append_file(std::FILE* from, std::FILE* append_to){
     int c = fgetc(from);
@@ -38,27 +35,4 @@ void append_file(std::FILE* from, std::FILE* append_to, const std::string& adden
 bool exists(std::string& file){ // use the file status library to check the existence of a file.
     struct stat file_info{};
     return !stat(file.c_str(), &file_info);
-}
-
-std::string get_indented_str(const std::string& str, int indent){
-    std::stringstream dump_stream;
-    dump_stream << std::setw(indent) << str << std::endl;
-    return dump_stream.str();
-}
-
-PrinterOp getPrinterOp(int i){
-    switch(i){
-        case 0:
-            return INIT_SPOOL;
-        case 1:
-            return END_SPOOL;
-        case 2:
-            return PRINT;
-        case 3:
-            return DUMP_SPOOL;
-        case 4:
-            return TERMINATE;
-        default:
-            return NOOP;
-    }
 }
